tests/wpt/web-platform-tests/html/webappapis/scripting/events/iframe-onload.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>iframe load event test</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <script>
      var iframe_loads = 0;
      function iframeOnLoad(t) {
        iframe_loads++;

        setTimeout(function() {
          t.step(function() {
            assert_true(iframe_loads === 1, "iframe.onload should be called exactly once.");
          });
          t.done();
        }, 100);
      }

      function startTest() {
        async_test(function(t) {
          t.step(function() {
            var iframe = document.createElement("iframe");
            iframe.onload = function() {
              iframeOnLoad(t);
            };
            document.body.appendChild(iframe);
          });
        }, "IFrameOnLoadFiredOnce");
      }
    </script>
  </head>
  <body onload="startTest()">
  </body>
</html>
